65 - Enhancement - Uso de DTOs para trafegar dados

+++ src/main/java/br/com/fiap/fastfood/api/adapters/driven/infrastructure/repository/adapter/OrderRepositoryAdapterImpl.java
@@ -6,9 +6,12 @@ import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.InvoiceMap
 import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapper;
 import br.com.fiap.fastfood.api.adapters.driven.infrastructure.mapper.OrderMapperImpl;
 import br.com.fiap.fastfood.api.adapters.driven.infrastructure.repository.order.OrderRepository;
+import br.com.fiap.fastfood.api.core.application.dto.invoice.InvoiceDTO;
+import br.com.fiap.fastfood.api.core.application.dto.order.OrderDTO;
 import br.com.fiap.fastfood.api.core.application.port.repository.OrderRepositoryPort;
 import br.com.fiap.fastfood.api.core.domain.model.invoice.Invoice;
 import br.com.fiap.fastfood.api.core.domain.model.order.Order;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -31,36 +34,33 @@ public class OrderRepositoryAdapterImpl implements OrderRepositoryPort {
   }
 
   @Override
-  public Optional<Order> findById(Long identifier) {
+  public Optional<OrderDTO> findById(Long identifier) {
     Optional<OrderEntity> orderEntityOptional = orderRepository.findById(identifier);
-    Optional<Order> orderOpt = orderEntityOptional.map(orderMapper::toDomain);
+    Optional<OrderDTO> orderOpt = orderEntityOptional.map(orderMapper::toDTO);
     if (orderEntityOptional.isPresent() && orderOpt.isPresent()) {
       OrderEntity entity = orderEntityOptional.get();
-      Order order = orderOpt.get();
-      List<Invoice> invoices = Optional.ofNullable(entity.getOrderInvoices()).orElse(Collections.emptyList()).stream().map(i -> {
-        Invoice invoice = invoiceMapper.toDomain(i);
-        invoiceMapper.mapStateImpl(i.getState(), invoice);
-        return invoice;
-      }).collect(
+      OrderDTO orderDTO = orderOpt.get();
+      List<InvoiceDTO> invoices = Optional.ofNullable(entity.getOrderInvoices()).orElse(new ArrayList<>()).stream().map(
+          invoiceMapper::toDTO).collect(
           Collectors.toList());
-      order.setInvoices(invoices);
-      order.setState(orderMapper.mapStateImpl(entity
-          .getState(), order));
+      orderDTO.setInvoices(invoices);
     }
     return orderOpt;
   }
 
   @Override
-  public Order save(Order data) {
+  public OrderDTO save(OrderDTO data) {
     OrderEntity entity = orderMapper.toEntity(data);
     OrderEntity persistedEntity = orderRepository.save(entity);
-    Order order = orderMapper.toDomain(persistedEntity);
-    order.setState(orderMapper.mapStateImpl(persistedEntity.getState(), order));
-    return order;
+    return orderMapper.toDTO(persistedEntity);
   }
 
   @Override
   public boolean delete(Long identifier) {
+    Optional<OrderEntity> orderEntityOpt = orderRepository.findById(identifier);
+    if (orderEntityOpt.isEmpty()) {
+      return false;
+    }
     orderRepository.deleteById(identifier);
     return true;
   }
